+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Improved codebase documentations.
 - Added `performanceMode` allowing developers to choose their preferred performance mode.
+- Include `IndicatorShape.none` allowing developers to hide face indicator.
 
 ## 0.0.9